+++ packages/cli/src/index.ts
@@ -107,6 +107,12 @@ const startCommand: yargs.CommandModule<unknown, StartArgs> = {
       .option('registry', {
         description: 'Registry with which the app is registered',
       })
+      .option('proxyUrl', {
+        default: process.env.APPSHELL_PROXY_URL || '',
+        type: 'string',
+        description: 'Proxy url for calls to get external resources',
+        global: true,
+      })
       .option('baseRegistry', {
         alias: 'a',
         default: [],
@@ -238,12 +244,6 @@ const generateEnvCommand: yargs.CommandModule<unknown, GenerateEnvArgs> = {
   // eslint-disable-next-line @typescript-eslint/no-shadow
   builder: (yargs) =>
     yargs
-      .option('env', {
-        alias: 'e',
-        default: '.env',
-        type: 'string',
-        description: 'The .env file to process',
-      })
       .option('outDir', {
         alias: 'o',
         default: 'dist',
@@ -268,13 +268,6 @@ const generateEnvCommand: yargs.CommandModule<unknown, GenerateEnvArgs> = {
         default: '__appshell_env__',
         type: 'string',
         description: 'Global variable name window[globalName] used in the output js',
-      })
-      .option('overwrite', {
-        alias: 'w',
-        default: false,
-        type: 'boolean',
-        description:
-          'If true, values in --env take precendent over those currently set in the environment',
       }),
   handler: generateEnvHandler,
 };
@@ -363,6 +356,13 @@ yargs(hideBin(process.argv))
   .option('apiKey', {
     alias: 'k',
     default: process.env.APPSHELL_API_KEY || config.apiKey || '',
+    type: 'string',
+    description: 'Api key to use for appshell registry',
+    global: true,
+  })
+  .option('apiKeyHeader', {
+    alias: 'h',
+    default: process.env.APPSHELL_API_KEY_HEADER || config.apiKeyHeader || 'x-api-key',
     type: 'string',
     description: 'Api key to use for appshell registry',
     global: true,