+++ packages/config/src/utils/loadJson.ts
@@ -9,6 +9,7 @@ type LoadJsonOptions = {
   insecure: boolean;
   target: string | RegExp;
   apiKey?: string;
+  apiKeyHeader?: string;
   proxyUrl?: string;
 };
 
@@ -17,12 +18,13 @@ const loadJson = async <T = Record<string, unknown>>(
   insecure: boolean,
   target: string | RegExp,
   apiKey?: string,
+  apiKeyHeader?: string,
   proxyUrl?: string,
 ): Promise<T[]> => {
   if (isValidUrl(jsonPathOrUrl)) {
     const url = proxyUrl ? `${proxyUrl}?target=${encodeURI(jsonPathOrUrl)}` : jsonPathOrUrl;
     console.log(`loading json from host ${url}`);
-    const headers = apiKey ? { apikey: apiKey } : {};
+    const headers = apiKey && apiKeyHeader ? { [apiKeyHeader]: apiKey } : {};
     const resp = await axios.get<T>(
       url,
       insecure
@@ -43,7 +45,9 @@ const loadJson = async <T = Record<string, unknown>>(
     if (stat.isDirectory()) {
       console.debug(`loading json from dir ${jsonPathOrUrl}`);
       const files = list(jsonPathOrUrl, 1, target);
-      const entries = files.map((file) => loadJson(file, insecure, target, apiKey, proxyUrl));
+      const entries = files.map((file) =>
+        loadJson(file, insecure, target, apiKey, apiKeyHeader, proxyUrl),
+      );
 
       const docs = await Promise.all(entries).then((items) => items.flat() as T[]);
 
@@ -67,14 +71,17 @@ export default async <T = Record<string, unknown>>(
     insecure: false,
     target: /\.json$/i,
     apiKey: undefined,
+    apiKeyHeader: undefined,
     proxyUrl: undefined,
   },
 ): Promise<T[]> => {
+  console.log(`loading json from ${jsonPathOrUrl}`, JSON.stringify(options, null, 2));
   const items = await loadJson<T[]>(
     jsonPathOrUrl,
     options.insecure,
     options.target,
     options.apiKey,
+    options.apiKeyHeader,
     options.proxyUrl,
   );